+++ src/ra/ww_audio.cc
@@ -490,7 +490,7 @@ void Sound_Effect(VocType voc, COORDINATE coord, int variation,
     }
   }
 
-  Sound_Effect(voc, volume, variation, pan_value, house);
+  Sound_Effect(voc, volume, variation, static_cast<short>(pan_value), house);
 }
 
 /***********************************************************************************************
@@ -865,7 +865,7 @@ void Speak_AI() {
       **	voice into the oldest buffer available.
       */
       if (speech == nullptr) {
-        _index = (_index + 1) % std::ssize(SpeechRecord);
+        _index = static_cast<int>((_index + 1) % std::ssize(SpeechRecord));
 
         auto name = std::filesystem::path(Speech[SpeakQueue])
                         .replace_extension(".AUD")
